型定義を移動。奥に押し込めた
